agents/mcp_tools/tool_manager.py
# multi_agent_system/agents/mcp_tools/tool_manager.py

# In this architecture, tools are primarily managed by:
# 1. Google ADK's FunctionTool: It wraps Python callables for LLM agents.
# 2. Python-A2A's AgentCard and Skill definitions: How A2A agents expose their functionalities.
# 3. protocols/a2a_protocol.py: Bridges A2A AgentCards into ADK FunctionTools.

# This file could be used if you had a generic "tool registry" that agents
# might pull from dynamically at runtime, but for now, ADK handles the orchestrator's
# tool management, and A2A handles individual agent tool exposure.

# For a concrete example, if we were to define a custom, non-A2A tool here, it would look like:
from google.adk.tools import FunctionTool
from loguru import logger

def example_internal_tool(text_input: str) -> str:
    """A simple internal tool that reverses a string."""
    logger.info(f"Executing example_internal_tool with: {text_input}")
    return text_input[::-1]

# This could then be added to an ADK agent's tools list:
# internal_reverse_tool = FunctionTool(example_internal_tool)

# This file is mostly illustrative for the MCP concept in a broader sense.
# Our current design uses A2A agents as the primary "tools" for the orchestrator.